#121 | refactor: Self Role Application Operations Have Been Separated with `RoleSelfApplicationService`
