+++ src/main/java/org/gelecekbilimde/scienceplatform/auth/service/RoleApplicationService.java
@@ -10,10 +10,6 @@ public interface RoleApplicationService {
 
 	Page<RoleApplication> findAll(List<RoleChangeRequestsFilter> filters, int page, int size);
 
-	void createAuthorApplication();
-
-	void createModeratorApplication();
-
 	void approve(String id);
 
 	void reject(String id);
