+++ src/main/java/org/gelecekbilimde/scienceplatform/auth/service/RoleSelfApplicationService.java
@@ -1,0 +1,9 @@
+package org.gelecekbilimde.scienceplatform.auth.service;
+
+public interface RoleSelfApplicationService {
+
+	void createAuthorApplication();
+
+	void createModeratorApplication();
+
+}
